add assert in MyTask
