+++ MyTask.cpp
@@ -17,6 +17,7 @@ bool is_less(const struct Airplane *l, const struct Airplane *r) {
 void merge_sort(struct Airplane *array, size_t size,
                 bool(isLess)(const struct Airplane *,
                              const struct Airplane *)) {
+  assert(array && (size > 0));
   int mid = size / 2;
   if (size % 2 == 1)
     mid++;
@@ -61,29 +62,4 @@ void merge_sort(struct Airplane *array, size_t size,
   free(c);
 }
 
-void task_function() {
-  printf("Input size:\n");
-  int size;
-  assert(scanf("%d", &size) == 1);
-  struct Airplane *airplane_array =
-      (struct Airplane *)malloc(size * sizeof(struct Airplane));
 
-  for (size_t i = 0; i < size; i++) {
-    // printf("Input %d-th Airplane:", size);
-    assert(scanf("%20s %20s %20s %d %d", airplane_array[i].brand,
-                 airplane_array[i].model, airplane_array[i].appointment,
-                 &airplane_array[i].crew_number,
-                 &airplane_array[i].range) == 5);
-  }
-  // Сортировка
-  merge_sort(airplane_array, size, is_less);
-  //
-  printf("Output data:\n");
-  for (size_t i = 0; i < size; i++) {
-    printf("%s %s %s %d %d\n", airplane_array[i].brand, airplane_array[i].model,
-           airplane_array[i].appointment, airplane_array[i].crew_number,
-           airplane_array[i].range);
-  }
-
-  free(airplane_array);
-}
